+++ packages/bpk-component-phone-input/src/BpkPhoneInput.js
@@ -58,6 +58,20 @@ type CommonProps = {
   valid: ?boolean,
 };
 
+// This function get the size value of the dialingCode to resize the field correctly
+const widthForDialingCode = (large, dialingCodes, selectedDialingCode) => {
+  // This sizeConstant is an average character size for each size of field
+  const averageLetterSize = large ? 8 : 6;
+  const foundDialingCode = dialingCodes.find(
+    e => e.code === selectedDialingCode,
+  );
+  if (foundDialingCode && foundDialingCode.description) {
+    // Here we calculate the width for the field (N.B 100 is an assumed padding value)
+    return averageLetterSize * foundDialingCode.description.length + 100;
+  }
+  return averageLetterSize + 100;
+};
+
 const BpkPhoneInput = (props: Props) => {
   const {
     id,
@@ -109,6 +123,9 @@ const BpkPhoneInput = (props: Props) => {
         wrapperClassName={getClassName(dialingCodeProps.wrapperClassName)}
         value={dialingCode}
         onChange={onDialingCodeChange}
+        style={{
+          width: `${widthForDialingCode(large, dialingCodes, dialingCode)}px`,
+        }}
       >
         {dialingCodes.map(({ code, description, ...extraDialingProps }) => (
           <option key={code} value={code} {...extraDialingProps}>